+++ src/forze_bigquery/kernel/_logger.py
@@ -1,0 +1,7 @@
+from forze.base.logging import Logger
+from forze_bigquery._logging import ForzeBigQueryLogger
+
+# ----------------------- #
+
+logger = Logger(ForzeBigQueryLogger.KERNEL)
+"""BigQuery kernel logger."""