+++ src/forze_clickhouse/kernel/_logger.py
@@ -1,0 +1,7 @@
+from forze.base.logging import Logger
+from forze_clickhouse._logging import ForzeClickHouseLogger
+
+# ----------------------- #
+
+logger = Logger(ForzeClickHouseLogger.KERNEL)
+"""ClickHouse kernel logger."""